Move projected world-event builders into a new world_events.rs seam
Pure, behavior-preserving extraction from main.rs toward ENG.md priority
#1 (tracked by #255): the 26 append_*_event RuntimeWorld methods that
construct EventViews, bump next_event_seq, and end at the
push_projected_event boundary move from main.rs into a new
world_events.rs seam as pub(crate) split-impl methods. main.rs shrinks
63391 -> 61903; the line ceiling lowers to lock in the win. Tests remain
in main.rs's root harness for this slice.
